Refactor Camunda package
- Join worker in single class.
- Add `UserTaskWorker` class as handler to `handler`` package
